refactor: replace management ui and remove smarttuple
