+++ app/JournalApp.tsx
@@ -39,7 +39,6 @@ export default function JournalApp() {
   const [showSaved, setShowSaved] = useState(false);
   const textareaRef = useRef<HTMLTextAreaElement>(null);
   const saveTimeoutRef = useRef<NodeJS.Timeout | null>(null);
-  const smartTupleTimeoutRef = useRef<NodeJS.Timeout | null>(null);
 
   const dateKey = format(selectedDate, 'yyyy-MM-dd');
 
@@ -76,16 +75,11 @@ export default function JournalApp() {
     setCurrentContent(entry?.content || '');
   }, [selectedDate, entries, dateKey]);
 
-  // Schedule smart tuple generation after user stops editing
-
   useEffect(() => {
     if (currentContent === (entries[dateKey]?.content || '')) return;
 
     if (saveTimeoutRef.current) {
       clearTimeout(saveTimeoutRef.current);
-    }
-    if (smartTupleTimeoutRef.current) {
-      clearTimeout(smartTupleTimeoutRef.current);
     }
 
     setIsSaving(true);
@@ -107,18 +101,11 @@ export default function JournalApp() {
       setShowSaved(true);
     }, 500);
 
-    smartTupleTimeoutRef.current = setTimeout(() => {
-      fetch('/api/smarttuple').catch(() => {});
-    }, 3000);
 
     return () => {
       if (saveTimeoutRef.current) {
         clearTimeout(saveTimeoutRef.current);
         saveTimeoutRef.current = null;
-      }
-      if (smartTupleTimeoutRef.current) {
-        clearTimeout(smartTupleTimeoutRef.current);
-        smartTupleTimeoutRef.current = null;
       }
       setIsSaving(false);
     };
